+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Improvements
  - Logging command is run in background
 
+### Updates
+ - Added `Logger` class for public use
+ - Changed `Residue::instance` to `Residue::init`
+
 ## [0.0.1]
  - Initial release